Refactor to package-based monorepo structure with enhanced file tasks
This commit restructures Worklift into a package-based monorepo with three
main packages:

**New Packages:**
- @worklift/core: Base Task class, project, target, and TaskScheduler
- @worklift/file-tasks: File and OS operation tasks
- @worklift/java-tasks: Java build tasks
- worklift: Meta-package that re-exports all functionality

**New File Tasks:**
- MoveTask: Move/rename files and directories
- CreateFileTask: Create files with content
- ZipTask: Create ZIP archives
- UnzipTask: Extract ZIP archives

**Existing Tasks (moved to packages):**
- CopyTask, DeleteTask, MkdirTask, ExecTask (now in @worklift/file-tasks)
- JavacTask, JarTask, JavaTask (now in @worklift/java-tasks)

**Configuration Updates:**
- Updated package.json with workspace configuration
- Added tsconfig path mappings for workspace packages
- Updated all examples to use new import structure
- Updated README with monorepo structure documentation

**Testing:**
All existing functionality has been preserved and tested. Examples run
successfully with the new package structure.
