+++ README.md
@@ -16,7 +16,35 @@ Worklift allows you to define build processes using TypeScript instead of XML. I
   - Targets can depend on specific targets in other projects
   - Automatic cyclic dependency detection
 - **Extensible**: Easy to add custom tasks and language-specific modules
-- **Built-in Tasks**: Common operations (copy, delete, mkdir, exec) and Java support (javac, jar, java)
+- **Package-based Architecture**: Modular monorepo structure with separate packages for core, file tasks, and Java tasks
+- **Built-in Tasks**: Common operations (copy, move, delete, mkdir, create file, zip, unzip, exec) and Java support (javac, jar, java)
+
+## Monorepo Structure
+
+Worklift uses a package-based monorepo structure:
+
+```
+worklift/
+├── packages/
+│   ├── core/              # @worklift/core - Base classes and project system
+│   ├── file-tasks/        # @worklift/file-tasks - File and OS operations
+│   ├── java-tasks/        # @worklift/java-tasks - Java build tasks
+│   └── worklift/          # worklift - Meta-package that re-exports everything
+├── examples/              # Example build scripts
+└── package.json           # Workspace configuration
+```
+
+You can import from individual packages or use the convenience `worklift` package:
+
+```typescript
+// Option 1: Import from the main package (recommended)
+import { project, CopyTask, JavacTask } from "worklift";
+
+// Option 2: Import from specific packages
+import { project } from "@worklift/core";
+import { CopyTask, MoveTask } from "@worklift/file-tasks";
+import { JavacTask } from "@worklift/java-tasks";
+```
 
 ## Installation
 
@@ -150,24 +178,40 @@ CopyTask.from("src/config")
 // Automatically tracks inputs/outputs for incremental builds
 ```
 
-## Available Tasks
+## Available Packages & Tasks
 
-### Common Tasks
+### @worklift/core
 
-Import from `worklift`:
+Core functionality including `Task` base class, `project`, `Target`, and the task scheduler.
+
+```typescript
+import { project, Task } from "@worklift/core";
+```
+
+### @worklift/file-tasks
+
+Common file and OS operations:
 
 - **CopyTask**: Copy files or directories
   - `CopyTask.from(path).to(destination).recursive(true).force(true)`
+- **MoveTask**: Move/rename files or directories
+  - `MoveTask.from(path).to(destination)`
 - **DeleteTask**: Delete files or directories
   - `DeleteTask.paths(...paths).recursive(true)`
 - **MkdirTask**: Create directories
   - `MkdirTask.paths(...paths)`
+- **CreateFileTask**: Create a file with content
+  - `CreateFileTask.path(file).content(data).encoding('utf-8')`
+- **ZipTask**: Create ZIP archives
+  - `ZipTask.from(path).to(file).recursive(true)`
+- **UnzipTask**: Extract ZIP archives
+  - `UnzipTask.file(path).to(dir).overwrite(true)`
 - **ExecTask**: Execute shell commands
   - `ExecTask.command(cmd).args([...]).cwd(dir).env({...})`
 
-### Java Tasks
+### @worklift/java-tasks
 
-Import from `worklift`:
+Java build tasks:
 
 - **JavacTask**: Compile Java source files
   - `JavacTask.sources(...files).destination(dir).classpath([...]).sourceVersion(ver).targetVersion(ver).encoding(enc)`
@@ -176,6 +220,14 @@ Import from `worklift`:
 - **JavaTask**: Run Java applications
   - `JavaTask.mainClass(className).classpath([...]).jvmArgs([...]).args([...])`
   - `JavaTask.jar(file).jvmArgs([...]).args([...])`
+
+### worklift (meta-package)
+
+Convenience package that re-exports all tasks and core functionality:
+
+```typescript
+import { project, CopyTask, MoveTask, JavacTask, ZipTask } from "worklift";
+```
 
 ## Example
