+++ packages/core/src/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @worklift/core - Core functionality for Worklift build tool
+ */
+
+export { Task } from "./Task.ts";
+export { project } from "./project.ts";
+export type { Project, Target, Dependency } from "./types.ts";
+export { TaskScheduler } from "./TaskScheduler.ts";
