+++ packages/core/src/project.ts
