+++ packages/file-tasks/src/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @worklift/file-tasks - Common file and OS tasks
+ */
+
+export { CopyTask } from "./CopyTask.ts";
+export { MoveTask } from "./MoveTask.ts";
+export { DeleteTask } from "./DeleteTask.ts";
+export { MkdirTask } from "./MkdirTask.ts";
+export { CreateFileTask } from "./CreateFileTask.ts";
+export { ZipTask } from "./ZipTask.ts";
+export { UnzipTask } from "./UnzipTask.ts";
+export { ExecTask } from "./ExecTask.ts";
