+++ packages/java-tasks/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@worklift/java-tasks",
+  "version": "0.1.0",
+  "description": "Java build tasks for Worklift build tool",
+  "module": "src/index.ts",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "dependencies": {
+    "@worklift/core": "workspace:*"
+  },
+  "devDependencies": {
+    "@types/node": "^22.0.0"
+  }
+}
